chore: move OpsLog
